src/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.13.1
frozendict==2.4.0
html5lib==1.1
idna==3.6
lxml==5.1.0
multitasking==0.0.11
numpy==1.26.4
pandas==2.2.0
peewee==3.17.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
virtualenv==20.25.0
webencodings==0.5.1
yfinance==0.2.36
